\name{@PKGNAME@-datasets}
\docType{data}

\alias{@PKGNAME@-datasets}
\alias{seqlengths}
@MASKSDATASETALIASES@

\title{@PKGNAME@ datasets}

\description{
  Data sets included in the @PKGNAME@ package in addition to the
  sequence data (and related to the sequence data).
}

\usage{
  data(seqlengths, package="@PKGNAME@")
  @MASKSDATASETUSAGES@
}

\author{@AUTHOR@}

\seealso{
  \link{@PKGNAME@},
  \link[BSgenome]{BSgenomeForge}
}

\examples{
  data(seqlengths, package="@PKGNAME@")
}

\keyword{package}
\keyword{data}
\keyword{datasets}
